fix(container): fix container bugs and add doc build
When running docs or build before shell commands in run.sh,
the next shell command would exit immideately after starting the
container.
This commit introduces a fix in the container entrypoint that
returns early in the first boot hook.
This allows the container script to create the container using
the run function without arguments - the entrypoint returns early
and the container script uses start and exec to attach itself,
running the users desired command.